// isoft-fe/src/app/models/tax_rates.model.ts
export class TaxRates {
    id: number;
    code: string;
    name: string;
    description: string;
    order: number;
    percent: number;
    type: number;
    debitCode: string;
    debitCodeName: string;
    creditCode: string;
    creditCodeName: string;
    debitFirstCode: string;
    debitSecondCode: string;
    creditFirstCode: string;
    creditSecondCode: string;
}